Working i2c code and MPR121 adaptation for orangepi
Adapted Adafruit MPR121 code for arduino so that it works with the
OrangePi

Tuned delays in i2c with the logic analyzer... there is a still a little
blip as SDA is pulsed down for 0.4us during switchover between 2 tx() ops
but trying to tune it out was causing i2c to fail everywhere else, so I'm
leaving it in.

Commented out printf statement in gpio_lib, which was causing my 600kHz
(before adding delays) code to run at 4kHz, so reads took forever because
I kept pulsing the pins between input and output.